Look-up value for JavaHome through Windows registry to load jvm.dll
This commit makes use of the nascent WinReg.jl functionality developed by @simonbyrne (not yet a registered package) to lookup the location of the JavaHome registry entry and then load jvm.dll based on the determined location of that path, instead of requiring that the user must have set a JAVA_HOME environment variable.  This commit also loads the copy of msvcr100.dll that is packaged with the JDK on Windows.  At present, testing has only been performed on Windows systems with JDK 1.8 installed.  Consequently, older and newer versions of jvm.dll might include a different version of the msvcr dll in their packaging that would need to be loaded prior to loading jvm.dll.  A more robust manner of detecting which version of msvc is packaged with the JDK that should be loaded prior to loading jvm.dll is likely needed.  Testing has been performed on the following combinations of Julia/Java/Windows.  64-bit Julia/Java/Windows.  32-bit Julia/Java/Windows.  32-bit Julia/Java with 64-bit Windows.