Use WinReg to set JAVA_HOME if needed
This commit makes use of WinReg to set the value of JAVA_HOME on Windows within the ENV environment variable dictionary if JAVA_HOME does not currently exist in that dictionary.  Having JAVA_HOME set in this manner does allow for proper loading of the MSVC runtime needed by jvm.dll even when a system environment variable for JAVA_HOME is not set.